move tour components to tour folder add simple map
