+++ package.json
@@ -20,6 +20,7 @@
     "geojson": "^0.5.0",
     "next": "^12.1.5",
     "next-auth": "^4.3.4",
+    "ol": "^6.14.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "reflect-metadata": "^0.1.13"
@@ -29,6 +30,7 @@
     "@testing-library/react": "^13.2.0",
     "@types/geojson": "^7946.0.8",
     "@types/node": "17.0.24",
+    "@types/openlayers": "^4.6.18",
     "@types/react": "^18.0.5",
     "@types/react-dom": "^18.0.1",
     "@types/react-test-renderer": "^18.0.0",
